move helper functions from client to utils
