+++ maestral/client.py
@@ -19,7 +19,7 @@ import functools
 import contextlib
 from datetime import datetime, timezone
 from typing import (
-    Callable, Union, Any, Type, Tuple, Iterator, List, TypeVar, Optional, TYPE_CHECKING
+    Callable, Union, Any, Type, Tuple, List, TypeVar, Optional, TYPE_CHECKING
 )
 
 # external imports
@@ -40,7 +40,7 @@ from maestral.errors import (
 )
 from maestral.config import MaestralState
 from maestral.constants import DROPBOX_APP_KEY
-from maestral.utils import natural_size
+from maestral.utils import natural_size, chunks, clamp
 
 if TYPE_CHECKING:
     from maestral.sync import SyncItem
@@ -814,17 +814,6 @@ class DropboxClient:
         results = self.flatten_results(results)
 
         return results
-
-
-# ==== helper functions ==================================================================
-
-def chunks(lst: List, n: int) -> Iterator[List]:
-    for i in range(0, len(lst), n):
-        yield lst[i:i + n]
-
-
-def clamp(n: _T, minn: _T, maxn: _T) -> _T:
-    return max(min(maxn, n), minn)
 
 
 # ==== conversion functions to generate error messages and types =========================
