+++ maestral/utils/__init__.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+from typing import List, Iterator, TypeVar
+
+
+_T = TypeVar('_T')
 
 
 def natural_size(num: float, unit: str = 'B', sep: bool = True) -> str:
@@ -19,3 +23,19 @@ def natural_size(num: float, unit: str = 'B', sep: bool = True) -> str:
 
     prefix = 'T'
     return f'{num:.1f}{sep_char}{prefix}{unit}'
+
+
+def chunks(lst: List, n: int) -> Iterator[List]:
+    """
+    Partitions an iterable into chunks of length ``n``.
+
+    :param lst: Iterable to partition.
+    :param n: Chunk size.
+    :returns: Iterator over chunks.
+    """
+    for i in range(0, len(lst), n):
+        yield lst[i:i + n]
+
+
+def clamp(n: _T, minn: _T, maxn: _T) -> _T:
+    return max(min(maxn, n), minn)
